changes:osdev/jmos/06-heap/src/ord_array.c fixed call to kfree (implicit casting to uint32_t)

+++ osdev/jmos/06-heap/src/ord_array.c
@@ -54,7 +54,7 @@ ordered_array_t place_ordered_array(void *addr, uint32_t max_size, lessthan_pred
  */
 void destroy_ordered_array(ordered_array_t *array)
 {
-    kfree(array->array);
+    kfree((uint32_t)array->array);
 }
 
 /*
